refactor(refactor-patrol): extract ReviewAgentRunner, stage run_cycle
Second architecture pass, behavior unchanged:

- Reviewer's agent spawn + usage recording move to ReviewAgentRunner; the
  agent_runner call(feature:, prompt:, output_path:, run_dir:) protocol is
  now the only seam, and Reviewer is pure per-feature orchestration.
- Commands::RefactorPatrol#run_cycle decomposes into named stages —
  resolve_project! / scoped_features / score_features / guard_theses /
  build_payload — with the module-wide (project_root, cfg, state) argument
  order.

Live-tested with a real-agent dry run through the new runner path (theses
parsed and admissible, caps flags applied, dry-run contract intact).
Autoreview: no correctness findings; argument-order inconsistency fixed.
